fix(storybook): extract parameters to reduce side-effects
