+++ .storybook/react/preview.js
@@ -1,4 +1,4 @@
 import '../styleConfig';
 import '../../packages/pharos/src/test/initComponents';
 
-export { parameters } from '../main/preview';
+export { parameters } from '../parameters';
